+++ src/cmd/switch.go
@@ -109,6 +109,7 @@ func determineSwitchData(repo *execute.OpenRepoResult, verbose bool) (*switchDat
 	repo.Config, exit, err = validate.Config(validate.ConfigArgs{
 		Backend:            &repo.Backend,
 		BranchesToValidate: localBranches,
+		FinalMessages:      repo.FinalMessages,
 		LocalBranches:      localBranches,
 		TestInputs:         &dialogTestInputs,
 		Unvalidated:        *repo.Config,
