+++ src/validate/config.go
@@ -2,21 +2,26 @@ package validate
 
 import (
 	"errors"
+	"fmt"
 
 	"github.com/git-town/git-town/v14/src/cli/dialog"
 	"github.com/git-town/git-town/v14/src/cli/dialog/components"
 	"github.com/git-town/git-town/v14/src/config"
+	"github.com/git-town/git-town/v14/src/config/configdomain"
+	"github.com/git-town/git-town/v14/src/config/gitconfig"
 	"github.com/git-town/git-town/v14/src/git"
 	"github.com/git-town/git-town/v14/src/git/gitdomain"
 	. "github.com/git-town/git-town/v14/src/gohacks/prelude"
+	"github.com/git-town/git-town/v14/src/gohacks/stringslice"
 	"github.com/git-town/git-town/v14/src/messages"
 )
 
 func Config(args ConfigArgs) (validatedResult *config.Config, aborted bool, err error) {
 	// enter and save main and perennials
 	var validatedMain gitdomain.LocalBranchName
-	var additionalPerennials gitdomain.LocalBranchNames
+	var validatedPerennials gitdomain.LocalBranchNames
 	if args.Unvalidated.Config.MainBranch.IsEmpty() {
+		var additionalPerennials gitdomain.LocalBranchNames
 		validatedMain, additionalPerennials, aborted, err = dialog.MainAndPerennials(dialog.MainAndPerennialsArgs{
 			DialogInputs:          args.TestInputs,
 			GetDefaultBranch:      args.Backend.DefaultBranch,
@@ -31,9 +36,9 @@ func Config(args ConfigArgs) (validatedResult *config.Config, aborted bool, err 
 		if err = args.Unvalidated.SetMainBranch(validatedMain); err != nil {
 			return validatedResult, false, err
 		}
+		validatedPerennials = append(args.Unvalidated.Config.PerennialBranches, additionalPerennials...)
 		if len(additionalPerennials) > 0 {
-			newPerennials := append(args.Unvalidated.Config.PerennialBranches, additionalPerennials...)
-			if err = args.Unvalidated.SetPerennialBranches(newPerennials); err != nil {
+			if err = args.Unvalidated.SetPerennialBranches(validatedPerennials); err != nil {
 				return validatedResult, false, err
 			}
 		}
@@ -75,6 +80,11 @@ func Config(args ConfigArgs) (validatedResult *config.Config, aborted bool, err 
 
 	// remove outdated lineage
 	err = args.Unvalidated.RemoveOutdatedConfiguration(args.LocalBranches)
+	if err != nil {
+		return validatedResult, abort, err
+	}
+	mainAndPerennials := append(validatedPerennials, validatedMain)
+	err = cleanupPerennialParentEntries(args.Unvalidated.Config.Lineage, mainAndPerennials, args.Unvalidated.GitConfig, args.FinalMessages)
 
 	return &args.Unvalidated, false, err
 }
@@ -82,7 +92,22 @@ func Config(args ConfigArgs) (validatedResult *config.Config, aborted bool, err 
 type ConfigArgs struct {
 	Backend            *git.BackendCommands
 	BranchesToValidate gitdomain.LocalBranchNames
+	FinalMessages      *stringslice.Collector
 	LocalBranches      gitdomain.LocalBranchNames
 	TestInputs         *components.TestInputs
 	Unvalidated        config.Config
+}
+
+// cleanupPerennialParentEntries removes outdated entries from the configuration.
+func cleanupPerennialParentEntries(lineage configdomain.Lineage, perennialBranches gitdomain.LocalBranchNames, access gitconfig.Access, finalMessages *stringslice.Collector) error {
+	for _, perennialBranch := range perennialBranches {
+		if lineage.Parent(perennialBranch).IsSome() {
+			if err := access.RemoveLocalConfigValue(gitconfig.NewParentKey(perennialBranch)); err != nil {
+				return err
+			}
+			lineage.RemoveBranch(perennialBranch)
+			finalMessages.Add(fmt.Sprintf(messages.PerennialBranchRemovedParentEntry, perennialBranch))
+		}
+	}
+	return nil
 }
